don't construct ex-info if we don't need it
Profiling an application that's parsing a few hundred dates at a
time, I found that 75% of the time was spent constructing `ex-info`s
(specifically, it looks like getting the stack is an expensive operation).

In many (most?) cases, we won't actually be throwing any error, so
constructing the `ex-info` every time seems wasteful. Instead, just
construct it when it's thrown.